# backend/openapi.yaml
openapi: 3.0.0
paths:
  /api/v4/ping:
    get:
      operationId: PingController_send
      summary: ""
      description: >-
        # Health check endpoint<br>


        This endpoint is a simple health check API designed to confirm that the
        server is operational.


        When accessed, it returns a straightforward response indicating that the
        service is up and running.
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ping"
  /api/v4/auth/login:
    get:
      operationId: AuthController_login
      parameters: []
      responses:
        "200":
          description: ""
        "302":
          description: Redirects to the AuthSch login page.
      tags: &a1
        - auth
  /api/v4/auth/callback:
    get:
      operationId: AuthController_oauthRedirect
      parameters:
        - name: code
          required: true
          in: query
          schema: {}
      responses:
        "200":
          description: ""
        "302":
          description: Redirects to the frontend with the JWT in the query string.
      tags: *a1
  /api/v4/auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
      tags: *a1
info:
  title: PÉK API
  description: Profiles and Groups
  version: v4
  contact:
    name: kir-dev
    url: https://kir-dev.hu
    email: hello@kir-dev.hu
tags: []
servers: []
components:
  schemas:
    Ping:
      type: object
      properties:
        ping:
          type: string
      required:
        - ping
    UserDto:
      type: object
      properties:
        name:
          type: string
      required:
        - name
externalDocs:
  description: Source Code (GitHub)
  url: https://github.com/kir-dev/pek-infinity